:recycle: refactor: refactor the api router to app route handlers
