+++ next.config.mjs
@@ -1,7 +1,6 @@
 import nextPWA from 'next-pwa';
 
 const isProd = process.env.NODE_ENV === 'production';
-const DEV_API_END_PORT_URL = process.env.DEV_API_END_PORT_URL || '';
 
 const withPWA = nextPWA({
   dest: 'public',
@@ -12,7 +11,6 @@ const withPWA = nextPWA({
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  pageExtensions: ['tsx', 'api.ts'],
   // not sure why antd-style cause multi ThemeProvider instance
   // So we need to transpile it to lib mode
   transpilePackages: ['@lobehub/ui', 'antd-style'],
@@ -25,22 +23,6 @@ const nextConfig = {
     return config;
   },
 
-  async rewrites() {
-    return [
-      {
-        source: '/api/openai/chat-dev',
-        destination: `${DEV_API_END_PORT_URL}/api/openai/chat`,
-      },
-      {
-        source: '/api/openai/models-dev',
-        destination: `${DEV_API_END_PORT_URL}/api/openai/models`,
-      },
-      {
-        source: '/api/plugins-dev',
-        destination: `${DEV_API_END_PORT_URL}/api/plugins`,
-      },
-    ];
-  },
   env: {
     AGENTS_INDEX_URL: process.env.AGENTS_INDEX_URL,
     PLUGINS_INDEX_URL: process.env.PLUGINS_INDEX_URL,
