+++ src/app/api/openai/createAzureOpenai.ts
