+++ src/app/api/openai/createOpenai.ts
