+++ src/app/api/openai/errorResponse.ts
