+++ src/app/api/openai/models/route.ts
@@ -1,10 +1,10 @@
 import { getOpenAIAuthFromRequest } from '@/const/fetch';
 
-import { createOpenai } from './createOpenai';
+import { createOpenai } from '../createOpenai';
 
 export const runtime = 'edge';
 
-export default async function handler(req: Request) {
+export const POST = async (req: Request) => {
   const { apiKey, endpoint } = getOpenAIAuthFromRequest(req);
 
   const openAI = createOpenai(apiKey, endpoint);
@@ -14,4 +14,4 @@ export default async function handler(req: Request) {
   const modelList = res.data.map((i) => i.id);
 
   return new Response(JSON.stringify(modelList));
-}
+};
